chatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
openai
qdrant-client
psycopg2-binary
python-dotenv
httpx
